requirements.txt
# Core
azure-eventhub>=5.11.0
azure-identity>=1.15.0
python-dotenv>=1.0.0

# Azure SQL
pyodbc>=5.0.0

# Azure Machine Learning
azure-ai-ml>=1.12.0
azureml-core>=1.54.0

# ML Training
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
joblib>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0

# Azure Blob Storage
azure-storage-blob>=12.19.0

# Azure Key Vault
azure-keyvault-secrets>=4.7.0

# Azure Data Factory
azure-mgmt-datafactory>=4.0.0

# Monitoring & Telemetry
opencensus-ext-azure>=1.1.0
opencensus-ext-logging>=0.1.0

# Visualization & Charts
matplotlib>=3.7.0
plotly>=5.18.0

# Web App
flask>=3.0.0
azure-functions>=1.21.0
APScheduler>=3.10.0

# Utilities
requests>=2.31.0
pytest>=8.0.0
pytest-cov>=5.0.0

# Log
python-json-logger>=4.0.0,<5.0.0

# Databricks (local deploy script only)
databricks-sdk>=0.20.0